直接购买-无购物车功能-查询默认的发票抬头

+++ src/main/java/com/acme/acmemall/service/IInvoiceTitleService.java
@@ -12,7 +12,7 @@ public interface IInvoiceTitleService {
 
     InvoiceTitleVo queryObject(Integer id);
 
-    InvoiceTitleVo queryDefaultByUserId(Integer userId);
+    InvoiceTitleVo queryDefaultByUserId(long userId);
 
     void update(InvoiceTitleVo entity);
